prevent adding some blocks inside Mini Cart Contents block
