+++ assets/js/blocks/cart-checkout/mini-cart-contents/inner-blocks/filled-mini-cart-contents-block/edit.tsx
@@ -14,7 +14,7 @@ import { useForcedLayout, getAllowedBlocks } from '../../../shared';
 
 export const Edit = ( { clientId }: { clientId: string } ): JSX.Element => {
 	const blockProps = useBlockProps();
-	const allowedBlocks = getAllowedBlocks( innerBlockAreas.EMPTY_MINI_CART );
+	const allowedBlocks = getAllowedBlocks( innerBlockAreas.FILLED_MINI_CART );
 	const { currentView } = useEditorContext();
 
 	const defaultTemplate = ( [
